requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
langchain-core==0.3.80
langchain-openai>=0.3.0
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
python-dotenv==1.2.1
pydantic==2.12.5
httpx==0.28.1
